Move tomato json conversion to its own module
