+++ aurora_cycler_manager/analysis.py
@@ -38,165 +38,6 @@ from aurora_cycler_manager.version import __url__, __version__
 
 warnings.filterwarnings("ignore", category=RuntimeWarning, message="All-NaN axis encountered")
 
-def convert_tomato_json(
-        snapshot_file: str,
-        output_hdf_file: str | None = None,
-        output_jsongz_file: str | None = None,
-        ) -> pd.DataFrame:
-    """Convert a raw json file from tomato to a pandas dataframe.
-
-    Args:
-        snapshot_file (str): path to the raw json file
-        output_hdf_file (str, optional): path to save the output hdf5 file
-        output_jsongz_file (str, optional): path to save the output json.gz file
-
-    Returns:
-        pd.DataFrame: DataFrame containing the cycling data
-
-    Columns in output DataFrame:
-    - uts: Unix time stamp in seconds
-    - V (V): Cell voltage in volts
-    - I (A): Current in amps
-    - loop_number: how many loops have been completed
-    - cycle_number: used if there is a loop of loops
-    - index: index of the method in the payload
-    - technique: code of technique using Biologic convention
-        100 = OCV, 101 = CA, 102 = CP, 103 = CV, 155 = CPLIMIT, 157 = CALIMIT,
-        -1 = Unknown
-
-    The dataframe is saved to 'cycling' key in the hdf5 file.
-    Metadata is added to the 'cycling' attributes in hdf5 file.
-    The metadata crucially includes json dumps of the job data and sample data
-    extracted from the database.
-
-    """
-    with open(snapshot_file, encoding="utf-8") as f:
-        input_dict = json.load(f)
-    n_steps = len(input_dict["steps"])
-    data = []
-    technique_code = {"NONE":0,"OCV":100,"CA":101,"CP":102,"CV":103,"CPLIMIT":155,"CALIMIT":157}
-    for i in range(n_steps):
-        step_data = input_dict["steps"][i]["data"]
-        step_dict = {
-            "uts" : [row["uts"] for row in step_data],
-            "V (V)" : [row["raw"]["Ewe"]["n"] for row in step_data],
-            "I (A)": [row["raw"]["I"]["n"] if "I" in row["raw"] else 0 for row in step_data],
-            "cycle_number": [row["raw"].get("cycle number", 0) for row in step_data],
-            "loop_number": [row["raw"].get("loop number", 0) for row in step_data],
-            "index" : [row["raw"].get("index", -1) for row in step_data],
-            "technique": [technique_code.get(row.get("raw", {}).get("technique"), -1) for row in step_data],
-        }
-        data.append(pd.DataFrame(step_dict))
-    data = pd.concat(data, ignore_index=True)
-    if output_hdf_file or output_jsongz_file:
-        # Try to get the job number from the snapshot file and add to metadata
-        try:
-            json_filename = os.path.basename(snapshot_file)
-            jobid = "".join(json_filename.split(".")[1:-1])
-            # look up jobid in the database
-            current_dir = os.path.dirname(os.path.abspath(__file__))
-            config_path = os.path.join(current_dir, "..", "config.json")
-            with open(config_path, encoding = "utf-8") as f:
-                config = json.load(f)
-            db_path = config["Database path"]
-            with sqlite3.connect(db_path) as conn:
-                conn.row_factory = sqlite3.Row
-                cursor = conn.cursor()
-                # Get all data about this job
-                cursor.execute("SELECT * FROM jobs WHERE `Job ID`=?", (jobid,))
-                job_data = dict(cursor.fetchone())
-                job_data["Payload"] = json.loads(job_data["Payload"])
-                sampleid = job_data["Sample ID"]
-                # Get all data about this sample
-                cursor.execute("SELECT * FROM samples WHERE `Sample ID`=?", (sampleid,))
-                sample_data = dict(cursor.fetchone())
-        except Exception as e:
-            print(f"Error getting job and sample data from database: {e}")
-            job_data = {}
-            sample_data = {}
-
-        # Create metadata
-        job_data["job_type"] = "tomato_0_2_biologic"
-        metadata = {
-            "provenance": {
-                "snapshot_file": snapshot_file,
-                "tomato_metadata": input_dict["metadata"],
-                "aurora_metadata": {
-                    "json_conversion": {
-                        "repo_url": __url__,
-                        "repo_version": __version__,
-                        "method": "analysis.py convert_tomato_json",
-                        "datetime": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
-                    },
-                },
-            },
-            "job_data": job_data,
-            "sample_data": sample_data,
-            "glossary": {
-                "uts": "Unix time stamp in seconds",
-                "V (V)": "Cell voltage in volts",
-                "I (A)": "Current across cell in amps",
-                "loop_number": "Number of loops completed from EC-lab loop technique",
-                "cycle_number": "Number of cycles within one technique from EC-lab",
-                "index": "index of the method in the payload, i.e. 0 for the first method, 1 for the second etc.",
-                "technique": "code of technique using definitions from MPG2 developer package",
-            },
-        }
-        if output_hdf_file:
-            folder = os.path.dirname(output_hdf_file)
-            if not folder:
-                folder = "."
-            if not os.path.exists(folder):
-                os.makedirs(folder)
-            data.to_hdf(
-                output_hdf_file,
-                key="cycling",
-                complib="blosc",
-                complevel=2,
-            )
-            with h5py.File(output_hdf_file, "a") as file:
-                if "cycling" in file:
-                    file["cycling"].attrs["metadata"] = json.dumps(metadata)
-                else:
-                    print("Dataset 'cycling' not found.")
-        if output_jsongz_file:
-            folder = os.path.dirname(output_jsongz_file)
-            if not folder:
-                folder = "."
-            if not os.path.exists(folder):
-                os.makedirs(folder)
-            full_data = {"data": data.to_dict(orient="list"), "metadata": metadata}
-            with gzip.open(output_jsongz_file, "wt", encoding="utf-8") as f:
-                json.dump(full_data, f)
-
-    return data
-
-def convert_all_tomato_jsons(
-    sampleid_contains: str = "",
-    ) -> None:
-    """Goes through all the raw json files in the snapshots folder and converts them to hdf5."""
-    current_dir = os.path.dirname(os.path.abspath(__file__))
-    config_path = os.path.join(current_dir, "..", "config.json")
-    with open(config_path, encoding = "utf-8") as f:
-        config = json.load(f)
-    raw_folder = config["Snapshots folder path"]
-    processed_folder = config["Processed snapshots folder path"]
-    for batch_folder in os.listdir(raw_folder):
-        for sample_folder in os.listdir(os.path.join(raw_folder, batch_folder)):
-            if sampleid_contains and sampleid_contains not in sample_folder:
-                continue
-            for snapshot_file in os.listdir(os.path.join(raw_folder, batch_folder, sample_folder)):
-                if snapshot_file.startswith("snapshot") and snapshot_file.endswith(".json"):
-                    output_folder = os.path.join(processed_folder,batch_folder,sample_folder)
-                    if not os.path.exists(output_folder):
-                        os.makedirs(output_folder)
-                    convert_tomato_json(
-                        os.path.join(raw_folder,batch_folder,sample_folder,snapshot_file),
-                        None,
-                        os.path.join(output_folder,snapshot_file.replace(".json",".json.gz")),
-                    )
-                    print(f"Converted {snapshot_file}")
-
 def combine_jobs(
         job_files: list[str],
 ) -> tuple[pd.DataFrame, dict]:
